Refactorization
The main functions_timeseries.py code was refactorized and splitted into 6 different scripts.

* Four of them contains the different models the main script will compare: auto_arima.py, holtwinter.py, lstm.py and var.py.
* One of them contains functions that will be used by different scripts, the helpers.py
* The last one is the main, the functions_timeseries.py. This one will use the other scripts in order to compare the models.

Server.py was also refactorized by improving the imports.

All this changes were tested before pushed and seemed to work fine.
